Project 8 QR code encoder  decoder/requirement.txt
pyzbar
Pillow
streamlit
